Les doubles sont de retour [18 mai 2022 à 21:58]

+++ includes/structure.h
@@ -6,7 +6,7 @@
 /*   By: cmariot <cmariot@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/05/01 22:38:52 by cmariot           #+#    #+#             */
-/*   Updated: 2022/05/18 20:44:04 by cmariot          ###   ########.fr       */
+/*   Updated: 2022/05/18 21:41:39 by cmariot          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -48,9 +48,9 @@ typedef struct s_mlx {
 
 // Structure couleur RGB
 typedef struct s_color {
-	float			r;
-	float			g;
-	float			b;
+	double			r;
+	double			g;
+	double			b;
 	int				trgb;
 }	t_color;
 
@@ -58,22 +58,22 @@ typedef struct s_color {
 typedef struct s_light {
 	t_m_print		print;
 	t_v3			position;
-	float			ratio;
+	double			ratio;
 }	t_light;
 
 // Lumiere ambiante
 typedef struct t_amb {
 	t_m_print		print;
 	t_color			color;
-	float			ratio;
-	float			ambient_constant;
+	double			ratio;
+	double			ambient_constant;
 }	t_amb;
 
 // Rayon
 typedef struct s_ray {
 	t_v3			position;
 	t_v3			direction;
-	float			t;
+	double			t;
 	t_v3			intersection;
 	t_v3			normale;
 }	t_ray;
@@ -84,12 +84,12 @@ typedef struct s_cam {
 	t_v3			position;
 	t_v3			direction;
 	t_ray			ray;
-	float			screen_width;
-	float			screen_height;
-	float			fov_horizontal;
-	float			constant_x;
-	float			constant_y;
-	float			constant_z;
+	double			screen_width;
+	double			screen_height;
+	double			fov_horizontal;
+	double			constant_x;
+	double			constant_y;
+	double			constant_z;
 }	t_cam;
 
 // Structure utilisée pour les sphere, cylindres et plans
@@ -101,8 +101,8 @@ typedef struct s_obj {
 	t_v3			position;
 	t_v3			direction;
 	t_color			color;
-	float			radius;
-	float			height;
+	double			radius;
+	double			height;
 	t_v3			ext1;
 	t_v3			ext2;
 	t_v3			axe;
