+++ includes/vector3.h
@@ -6,7 +6,7 @@
 /*   By: cmariot <cmariot@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/05/01 22:25:15 by cmariot           #+#    #+#             */
-/*   Updated: 2022/05/18 18:59:50 by cmariot          ###   ########.fr       */
+/*   Updated: 2022/05/18 21:41:44 by cmariot          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -15,24 +15,24 @@
 
 typedef struct s_v3
 {
-	float			x;
-	float			y;
-	float			z;
-	float			a;
+	double			x;
+	double			y;
+	double			z;
+	double			a;
 }	t_v3;
 
-t_v3		new_vector(const float x, const float y, const float z);
+t_v3		new_vector(const double x, const double y, const double z);
 
 t_v3		add_vector(const t_v3 a, const t_v3 b);
 t_v3		sub_vector(const t_v3 a, const t_v3 b);
-t_v3		div_vector(const t_v3 a, const float b);
-t_v3		mul_vector(const t_v3 a, const float b);
+t_v3		div_vector(const t_v3 a, const double b);
+t_v3		mul_vector(const t_v3 a, const double b);
 
 t_v3		normalize(const t_v3 a);
 t_v3		cross_product(const t_v3 a, const t_v3 b);
-float		norm_square(const t_v3 a);
-float		norm(const t_v3 a);
-float		scalar_product(const t_v3 a, const t_v3 b);
-float		length(const t_v3 point_a, const t_v3 point_b);
+double		norm_square(const t_v3 a);
+double		norm(const t_v3 a);
+double		scalar_product(const t_v3 a, const t_v3 b);
+double		length(const t_v3 point_a, const t_v3 point_b);
 
 #endif
