+++ libft/includes/libft.h
@@ -6,7 +6,7 @@
 /*   By: cmariot <cmariot@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/04/10 13:34:23 by cmariot           #+#    #+#             */
-/*   Updated: 2022/05/18 18:25:53 by cmariot          ###   ########.fr       */
+/*   Updated: 2022/05/18 21:42:23 by cmariot          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -96,7 +96,7 @@ char		*ft_strtrim(char const *s1, char const *set);
 char		*ft_substr(char const *s, unsigned int start, size_t len);
 int			ft_tolower(int c);
 int			ft_toupper(int c);
-int			trgb_color(int t, float r, float g, float b);
+int			trgb_color(int t, double r, double g, double b);
 
 /* GET NEXT LINE */
 char		*get_next_line(int fd);
